rename rayjobsubmitter to rayjob-subitter
