+++ .buildkite/build-load-light-weight-submitter.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
 
 # This script is used to build and load the light-weight submitter image into the kind cluster.
-IMG=kuberay/submitter:nightly make docker-image-rayjobsubmitter &&
+IMG=kuberay/submitter:nightly make docker-image-rayjob-submitter &&
     kind load docker-image kuberay/submitter:nightly
