+++ .buildkite/test-e2e.yml
@@ -141,7 +141,7 @@
     - echo "--- END:Apiserver e2e (nightly operator) tests finished"
 
 
-- label: 'Test RayJobSubmitter E2E (nightly operator)'
+- label: 'Test RayJob Submitter E2E (nightly operator)'
   instance_size: large
   image: golang:1.24-bookworm
   commands:
@@ -155,10 +155,10 @@
     - kubectl wait --timeout=90s --for=condition=Available=true deployment kuberay-operator
 
     # Run e2e tests and print KubeRay api server logs if tests fail
-    - echo "--- START:Running e2e (nightly operator) RayJobSubmitter tests"
+    - echo "--- START:Running e2e (nightly operator) RayJob Submitter tests"
     - if [ -n "${KUBERAY_TEST_RAY_IMAGE}"]; then echo "Using Ray Image ${KUBERAY_TEST_RAY_IMAGE}"; fi
     - set -o pipefail
     - mkdir -p "$(pwd)/tmp" && export KUBERAY_TEST_OUTPUT_DIR=$(pwd)/tmp
     - echo "KUBERAY_TEST_OUTPUT_DIR=$$KUBERAY_TEST_OUTPUT_DIR"
     - KUBERAY_TEST_TIMEOUT_SHORT=1m KUBERAY_TEST_TIMEOUT_MEDIUM=5m KUBERAY_TEST_TIMEOUT_LONG=10m go test -timeout 30m -v ./test/e2erayjobsubmitter 2>&1 | awk -f ../.buildkite/format.awk | tee $$KUBERAY_TEST_OUTPUT_DIR/gotest.log || (kubectl logs --tail -1 -l app.kubernetes.io/name=kuberay | tee $$KUBERAY_TEST_OUTPUT_DIR/kuberay-operator.log && cd $$KUBERAY_TEST_OUTPUT_DIR && find . -name "*.log" | tar -cf /artifact-mount/e2e-log.tar -T - && exit 1)
-    - echo "--- END:RayJobSubmitter E2E (nightly operator) tests finished"
+    - echo "--- END:RayJob Submitter E2E (nightly operator) tests finished"
